+++ src/components/auth/form/FindPassword.tsx
@@ -1,8 +1,8 @@
 import Button from '@/components/common/Button';
 import { useAuthSetQuery } from '@/hooks/query/auth/useAuthSetQuery';
 import { SubmitHandler, useForm } from 'react-hook-form';
+import Input from '../../common/Input';
 import styles from '../styles/Auth.module.css';
-import Input from './Input';
 
 const FindPassword = () => {
   const { register, handleSubmit } = useForm<Inputs>();
