+++ src/components/auth/form/Signup.tsx
@@ -4,9 +4,9 @@ import { isEmptyObject } from '@/shared/helper';
 import { Fragment } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { FaCheck as Check } from 'react-icons/fa6';
-import { signupInput } from '../../../data/inputData';
+import { signupInput } from '../../../data/input-props';
+import Input from '../../common/Input';
 import styles from '../styles/Auth.module.css';
-import Input from './Input';
 
 const MAX_BNO_LENGTH = 10;
 const SUCCESS_MSG = '사업자등록번호가 인증되었습니다.';
